research_agent/requirements.txt
httpx==0.28.1
pydantic==2.11.9
pydantic_ai==1.0.12
pydantic_settings==2.11.0
python-dotenv==1.1.1
streamlit==1.46.1
openai==1.56.0
